Refactor inputs on ScenarioFormatter into subcomponent.
